backend/requirements.txt
Kaggle
pandas
python-dotenv
